feat: label removal function
