+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/logo-a.png" />
+    <link rel="icon" type="image/svg+xml" href="/logo-active.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>医学数据标注平台</title>
   </head>
