+++ src/style.less
@@ -15,4 +15,20 @@
     user-select: none;
     padding: 4px 10px;
     border-radius: 20px;
+    position: relative;
+    pointer-events: none; // 禁止点击事件
+}
+
+// 由于surroundContents会清空span的子元素，故只能用伪元素来设置图标
+.onselect::after {
+    content: '';
+    display: inline-block;
+    background-image: url('./assets/close.svg');
+    background-repeat: no-repeat;
+    background-size: 20px 20px;
+    height: 20px;
+    width: 20px;
+    position: relative;
+    top: 5px;
+    pointer-events: auto; // 实现只在伪元素上才有的单击效果
 }
